Range key added to lookups

+++ src/Business/BusinessRequest.cs
@@ -4,7 +4,7 @@ namespace Business
 {
     public class BusinessRequest
     {
-        public int RequestingUserId { get; set; }
+        public long RequestingUserId { get; set; }
         public DateTime RequestedAt { get; set; }
     }
 }
